requirements.txt
pandas
psycopg2-binary
requests
fastapi
uvicorn[standard]
httpx
python-dotenv
openai
beautifulsoup4
lxml
